pyproject.toml
[tool.poetry]
name = "fastapi-example"
version = "0.1.0"
description = ""
authors = ["ArtemMor <moruginartem99@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
motor = "^3.4.0"
uvicorn = "^0.29.0"
ipython = "^8.22.2"
pytest = "^8.1.1"
faker = "^24.4.0"
pytest-asyncio = "^0.23.6"
punq = "^0.7.0"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
aiokafka = "^0.10.0"
orjson = "^3.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
